+++ packages/attention/src/component.tsx
@@ -6,12 +6,36 @@ import {
   useRecompute as recompute,
 } from '@warp-ds/core/attention'
 import { attention as ccAttention } from '@warp-ds/css/component-classes'
-import { ArrowProps, AttentionProps } from './props'
+import { ArrowProps, AttentionProps, AttentionVariants } from './props'
 import { i18n } from '@lingui/core'
 import { messages as nbMessages } from './locales/nb/messages.mjs'
 import { messages as enMessages } from './locales/en/messages.mjs'
 import { messages as fiMessages } from './locales/fi/messages.mjs'
 import { activateI18n } from '../../i18n'
+import { IconClose16 } from "@warp-ds/icons/react";
+
+const variantClasses = {
+  callout: {
+    wrapper: ccAttention.callout,
+    arrow: ccAttention.arrowCallout
+  },
+  highlight: {
+    wrapper: ccAttention.highlight,
+    arrow: ccAttention.arrowHighlight
+  },
+  tooltip: {
+    wrapper: ccAttention.tooltip,
+    arrow: ccAttention.arrowTooltip
+  },
+  popover: {
+    wrapper: ccAttention.popover,
+    arrow: ccAttention.arrowPopover
+  },
+}
+
+const getVariant = (variantProps: AttentionVariants) => {
+  return Object.keys(variantClasses).find(b => !!variantProps[b]) || '';
+};
 
 export function Attention(props: AttentionProps) {
   const {
@@ -23,16 +47,17 @@ export function Attention(props: AttentionProps) {
     placement,
     targetEl,
     className,
+    canClose,
+    onDismiss,
     ...rest
   } = props
 
   activateI18n(enMessages, nbMessages, fiMessages)
 
-  const wrapperClasses = classNames(ccAttention.base, {
-    [ccAttention.tooltip]: props.tooltip,
-    [ccAttention.callout]: props.callout,
-    [ccAttention.popover]: props.popover,
-  })
+  const wrapperClasses = classNames(
+    ccAttention.base,
+    variantClasses[getVariant(rest)].wrapper
+  );
 
   const [actualDirection, setActualDirection] = useState(placement)
   // Don't show attention element before its position is computed on first render
@@ -80,28 +105,28 @@ export function Attention(props: AttentionProps) {
     switch (opposites[actualDirection]) {
       case 'top':
         return i18n._({
-          id: 'attention.pointingUp',
+          id: 'attention.aria.pointingUp',
           message: 'pointing up',
           comment:
             'Default screenreader message for top direction in the attention component',
         })
       case 'right':
         return i18n._({
-          id: 'attention.pointingRight',
+          id: 'attention.aria.pointingRight',
           message: 'pointing right',
           comment:
             'Default screenreader message for right direction in the attention component',
         })
       case 'bottom':
         return i18n._({
-          id: 'attention.pointingDown',
+          id: 'attention.aria.pointingDown',
           message: 'pointing down',
           comment:
             'Default screenreader message for bottom direction in the attention component',
         })
       case 'left':
         return i18n._({
-          id: 'attention.pointingLeft',
+          id: 'attention.aria.pointingLeft',
           message: 'pointing left',
           comment:
             'Default screenreader message for left direction in the attention component',
@@ -116,24 +141,31 @@ export function Attention(props: AttentionProps) {
     switch (true) {
       case props.tooltip:
         return i18n._({
-          id: 'attention.tooltip',
+          id: 'attention.aria.tooltip',
           message: 'tooltip',
           comment:
             'Default screenreader message for tooltip in the attention component',
         })
       case props.callout:
         return i18n._({
-          id: 'attention.callout',
+          id: 'attention.aria.callout',
           message: 'callout speech bubble',
           comment:
             'Default screenreader message for callout speech bubble in the attention component',
         })
       case props.popover:
         return i18n._({
-          id: 'attention.popover',
+          id: 'attention.aria.popover',
           message: 'popover speech bubble',
           comment:
             'Default screenreader message for popover speech bubble in the attention component',
+        })
+      case props.highlight:
+        return i18n._({
+          id: 'attention.aria.highlight',
+          message: 'highlighted speech bubble',
+          comment:
+            'Default screenreader message for highlighted speech bubble in the attention component',
         })
       default:
         return ''
@@ -164,7 +196,6 @@ export function Attention(props: AttentionProps) {
 
   return (
     <div
-      tabIndex={0}
       className={classNames(
         {
           [ccAttention.notCallout]: !props.callout,
@@ -179,11 +210,36 @@ export function Attention(props: AttentionProps) {
         role={props.role === '' ? undefined : (props.tooltip ? 'tooltip' : 'img')}
         aria-label={ariaLabel === '' ? undefined : ariaLabel ?? defaultAriaLabel()}
         className={wrapperClasses}
+        id={props.id}
       >
         {!props.noArrow && (
           <Arrow {...props} ref={arrowRef} direction={placement} />
         )}
-        <div className={ccAttention.content}>{props.children}</div>
+        <div className={ccAttention.content}>
+          {props.children}
+        </div>
+        {canClose && (
+          <button
+            type="button"
+            aria-label={i18n._(
+              /*i18n*/ {
+                id: "attention.aria.close",
+                message: "Close",
+                comment: "Aria label for the close button in attention",
+              }
+            )}
+            onClick={onDismiss}
+            onKeyDown={(event) => {
+              if (!props.onDismiss) return;
+              if (event.key === "Escape") {
+                props.onDismiss();
+              }
+            }}
+            className={ccAttention.closeBtn}
+          >
+            <IconClose16 />
+          </button>
+        )}
       </div>
     </div>
   )
@@ -195,18 +251,13 @@ const arrowDirectionClassname = (dir: string) => {
   return `arrowDirection${direction}`
 }
 
-const Arrow = forwardRef<HTMLDivElement, ArrowProps>((props, ref) => {
-  const { callout, direction, popover, tooltip } = props
+const Arrow = forwardRef<HTMLDivElement, ArrowProps>(({ direction, ...rest }, ref) => {
   const arrowDirection = opposites[direction]
 
   const arrowClasses = classNames(
     ccAttention.arrowBase,
     ccAttention[arrowDirectionClassname(arrowDirection)],
-    {
-      [ccAttention.arrowTooltip]: tooltip,
-      [ccAttention.arrowCallout]: callout,
-      [ccAttention.arrowPopover]: popover,
-    }
+    variantClasses[getVariant(rest)].arrow
   )
 
   return (