+++ packages/attention/stories/Attention.stories.tsx
@@ -10,10 +10,10 @@ export function Callout() {
   return (
     <div className='flex items-center'>
       <Box info>
-        <h1 aria-details="callout-bubbletext">I am a box full of info</h1>
+        <h1 aria-details='callout-bubbletext'>I am a box full of info</h1>
       </Box>
       <Attention callout placement='right' isShowing={true}>
-        <p id="callout-bubbletext" style={{ width: 200 }}>
+        <p id='callout-bubbletext' style={{ width: 200 }}>
           I'm a callout because that box over there is new or something
         </p>
       </Attention>
@@ -24,7 +24,7 @@ export function CalloutResettingRoleAndAriaLabel() {
   return (
     <div className='flex items-center'>
       <Box info>
-        <h1 aria-details="callout-reset-bubbletext">I am a box full of info</h1>
+        <h1 aria-details='callout-reset-bubbletext'>I am a box full of info</h1>
       </Box>
       <Attention
         callout
@@ -33,10 +33,89 @@ export function CalloutResettingRoleAndAriaLabel() {
         placement='right'
         isShowing={true}
       >
-        <p id="callout-reset-bubbletext" role="img" style={{ width: 200 }}>
-          I'm a callout speech bubble with resetted role and aria-label attributes pointing left.
+        <p id='callout-reset-bubbletext' role='img' style={{ width: 200 }}>
+          I'm a callout speech bubble with resetted role and aria-label
+          attributes pointing left.
         </p>
       </Attention>
+    </div>
+  )
+}
+
+export function Highlight() {
+  const [show, setShow] = React.useState(false)
+  const targetEl = React.useRef()
+
+  return (
+    <div className='flex flex-col justify-between h-[200]'>
+      <Button
+        small
+        aria-expanded={show}
+        aria-controls='highlight-attention-example'
+        utility
+        onClick={() => setShow(!show)}
+        className='w-max mb-0'
+      >
+        Show an onboarding hint
+      </Button>
+      <div>
+        <div ref={targetEl} className="w-2/3">
+          <Box info>
+            <h1>I am a box full of info</h1>
+          </Box>
+        </div>
+        <Attention 
+          highlight
+          placement='top'
+          isShowing={show}
+          targetEl={targetEl}
+          id='highlight-attention-example'
+        >
+          <p>
+            I'm a highlight because that box over there is new or something
+          </p>
+        </Attention>
+      </div>
+    </div>
+  )
+}
+
+export function DismissibleHighlight() {
+  const [show, setShow] = React.useState(false)
+  const targetEl = React.useRef()
+
+  return (
+    <div className='flex flex-col justify-between h-[200]'>
+      <Button
+        small
+        aria-expanded={show}
+        aria-controls='highlight-attention-example'
+        utility
+        onClick={() => setShow(!show)}
+        className='w-max mb-0'
+      >
+        Show an onboarding hint
+      </Button>
+      <div>
+        <div ref={targetEl} className="w-2/3">
+          <Box info>
+            <h1>I am a box full of info</h1>
+          </Box>
+        </div>
+        <Attention
+          highlight
+          canClose
+          onDismiss={() => setShow(false)}
+          placement='top'
+          isShowing={show}
+          targetEl={targetEl}
+          id='highlight-attention-example'
+        >
+          <p>
+            I'm a highlight that can dismiss itself
+          </p>
+        </Attention>
+      </div>
     </div>
   )
 }
@@ -56,16 +135,19 @@ export function Tooltip() {
       <h1
         className='w-max mb-0'
         ref={targetEl}
-        aria-details='tooltip-bubbletext'
-        aria-hidden={!show}
-        onMouseEnter={() => setShow(true)}
-        onMouseLeave={() => setShow(false)}
-        onFocus={() => setShow(true)}
-        onBlur={() => setShow(false)}
-        onKeyDown={handleKeyDown}
-        tabIndex={0}
       >
-        hover this for useless info
+        <Button
+          aria-describedby='tooltip-bubbletext'
+          aria-expanded='true'
+          type='button'
+          className='bg-transparent'
+          onMouseEnter={() => setShow(true)}
+          onMouseLeave={() => setShow(false)}
+          onFocus={() => setShow(true)}
+          onBlur={() => setShow(false)}
+        >
+          hover this for useless info
+        </Button>
       </h1>
       <Attention
         tooltip
@@ -73,7 +155,7 @@ export function Tooltip() {
         targetEl={targetEl}
         isShowing={show}
       >
-        <p id="tooltip-bubbletext">lol i am a tooltip</p>
+        <p id='tooltip-bubbletext'>lol i am a tooltip</p>
       </Attention>
     </div>
   )
@@ -101,8 +183,7 @@ export function Popover() {
       <Button
         small
         aria-expanded={show}
-        aria-controls='pop-over-attention-example'
-        aria-details="pop-over-bubbletext"
+        aria-controls='popover-attention-example'
         utility
         onClick={() => setShow(!show)}
         className='w-max mb-0'
@@ -116,7 +197,7 @@ export function Popover() {
         targetEl={targetEl}
         isShowing={show}
       >
-        <ul className='bg-white w-full text-center' id="pop-over-bubbletext">
+        <ul className='bg-white w-full text-center'>
           <li
             tabIndex={0}
             className='p-2 px-40 hover:bg-blue-200 hover:text-blue-500 cursor-pointer'